Split the mutation-testing contract into its own module
The mutation-testing contract had accumulated inside the general workflows
helper, leaving `workflows.py` and `test_helpers.py` as the two largest
files in the test tree and mixing an unrelated shell parser in with the
GitHub Actions assertions.

Move the mutation contract into `tooling_contracts/mutation.py`: the
`_MUTATION_*` constants, the setup-commands shell parser
(`_split_shell_commands`, `_apt_install_packages`,
`_extract_apt_install_packages`), and `_assert_mutation_workflow_contracts`.
Point the orchestration entrypoint at the new module and drop the
now-unused `shlex` import from `workflows.py`. Move the matching
apt-install and mutation-schedule tests into `tests/test_mutation_contracts.py`,
alongside the strategies they use.

The move is behaviour-preserving: the suite still reports 77 passed and 1
skipped. Every touched file now sits well under 400 logical lines
(workflows 175, mutation 70, test_helpers 247, test_mutation_contracts 57).

Also trim the `_mutation_workflow_with_schedule` docstring to a single-line
summary, matching the sibling private helpers.
